pyproject.toml
[tool.poetry]
name = "alumni-portal"
version = "0.1.0"
description = ""
authors = ["Tariq Jamal A <tariqjamal4267324@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-crispy-forms = "^2.3"
django-social-share = "^2.3.0"
django-import-export = "^4.1.1"
pillow = "^10.4.0"
django-tinymce = "^4.1.0"
mysqlclient = "^2.2.4"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
docformatter = "^1.7.5"
djlint = "^1.35.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[tool.docformatter]
recursive = true
wrap-summaries = 99
wrap-descriptions = 99

[tool.djlint]
blank_line_after_tag="load,extends,include,block,endblock,render_form_field"
blank_line_before_tag="load,extends,include,block,endblock,render_form_field"
close_void_tags=true
exclude=".venv,venv,env"
format_attribute_template_tags=true
format_css=true
format_js=true
indent_size=4
max_line_length=120